requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
dotmap

# ENAS-pytorch requirements
ipdb
tqdm
scipy
imageio
pygraphviz
tensorboardX

# OTHER
torch
torchvision